+++ vehicle/src/Vehicle/Compile.hs
@@ -9,6 +9,7 @@ where
 
 import Control.Monad.Writer (MonadWriter (..), WriterT (..))
 import Vehicle.Backend.Agda
+import Vehicle.Backend.Isabelle
 import Vehicle.Backend.Loss (convertToLossTensors)
 import Vehicle.Backend.Loss.JSON
 import Vehicle.Backend.Prelude
@@ -130,6 +131,10 @@ compileToITP ITPOptions {..} typedProg = do
         let rocqOptions = RocqOptions outputFile moduleName
         rocqCode <- compileProgToRocq decProg rocqOptions
         writeRocqFile outputFile rocqCode
+      Isabelle -> do
+        let isabelleOptions = IsabelleOptions outputFile moduleName
+        isabelleCode <- compileProgToIsabelle decProg isabelleOptions
+        writeIsabelleFile outputFile isabelleCode
 
 compileToLossFunction ::
   forall m.